# frontend/openapi/apis.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI-Redux Posts App",
        "summary": "OpenAPI Specs for a FastAPI backend to serve a Redux-powered NextJS frontend for a simple Posts app.",
        "version": "0.1.0"
    },
    "paths": {
        "/auth/sign-up": {
            "post": {
                "tags": ["auth"],
                "summary": "Create User",
                "operationId": "create_user_auth_sign_up_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/UserCreate" }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/User" }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "tags": ["auth"],
                "summary": "Authenticate User",
                "operationId": "authenticate_user_auth_sign_in_post",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_authenticate_user_auth_sign_in_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/users/": {
            "get": {
                "tags": ["users"],
                "summary": "Get Authenticated User",
                "operationId": "get_authenticated_user_users__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/User" }
                            }
                        }
                    }
                },
                "security": [{ "OAuth2PasswordBearer": [] }]
            },
            "delete": {
                "tags": ["users"],
                "summary": "Delete Authenticated User",
                "operationId": "delete_authenticated_user_users__delete",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    }
                },
                "security": [{ "OAuth2PasswordBearer": [] }]
            }
        },
        "/users/{user_id}/posts": {
            "get": {
                "tags": ["users"],
                "summary": "Get User Posts",
                "operationId": "get_user_posts_users__user_id__posts_get",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "User Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Post" },
                                    "title": "Response Get User Posts Users  User Id  Posts Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{post_id}": {
            "get": {
                "tags": ["posts"],
                "summary": "Get Post",
                "operationId": "get_post_posts__post_id__get",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Post Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Post" }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["posts"],
                "summary": "Edit Post",
                "operationId": "edit_post_posts__post_id__put",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Post Id" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/PostCreate" }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Post" }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["posts"],
                "summary": "Delete Post",
                "operationId": "delete_post_posts__post_id__delete",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Post Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{post_id}/likes": {
            "get": {
                "tags": ["posts"],
                "summary": "Get Likes On Post",
                "operationId": "get_likes_on_post_posts__post_id__likes_get",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Post Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Like" },
                                    "title": "Response Get Likes On Post Posts  Post Id  Likes Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{post_id}/comments": {
            "get": {
                "tags": ["posts"],
                "summary": "Get Comments On Post",
                "operationId": "get_comments_on_post_posts__post_id__comments_get",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Post Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Comment" },
                                    "title": "Response Get Comments On Post Posts  Post Id  Comments Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/posts/": {
            "post": {
                "tags": ["posts"],
                "summary": "Create Post",
                "operationId": "create_post_posts__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/PostCreate" }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Post" }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                },
                "security": [{ "OAuth2PasswordBearer": [] }]
            }
        },
        "/comments/": {
            "post": {
                "tags": ["comments"],
                "summary": "Create Comment",
                "operationId": "create_comment_comments__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/CommentCreate" }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Comment" }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                },
                "security": [{ "OAuth2PasswordBearer": [] }]
            }
        },
        "/comments/{comment_id}": {
            "put": {
                "tags": ["comments"],
                "summary": "Edit Comment",
                "operationId": "edit_comment_comments__comment_id__put",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Comment Id" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/CommentCreate" }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Comment" }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/likes/post/{post_id}": {
            "post": {
                "tags": ["likes"],
                "summary": "Like A Post",
                "operationId": "like_a_post_likes_post__post_id__post",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Post Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/Like" }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["likes"],
                "summary": "Remove Like From A Post",
                "operationId": "remove_like_from_a_post_likes_post__post_id__delete",
                "security": [{ "OAuth2PasswordBearer": [] }],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Post Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "summary": "Ping",
                "operationId": "ping__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_authenticate_user_auth_sign_in_post": {
                "properties": {
                    "grant_type": {
                        "anyOf": [{ "type": "string", "pattern": "password" }, { "type": "null" }],
                        "title": "Grant Type"
                    },
                    "username": { "type": "string", "title": "Username" },
                    "password": { "type": "string", "title": "Password" },
                    "scope": { "type": "string", "title": "Scope", "default": "" },
                    "client_id": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Client Id"
                    },
                    "client_secret": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Client Secret"
                    }
                },
                "type": "object",
                "required": ["username", "password"],
                "title": "Body_authenticate_user_auth_sign_in_post"
            },
            "Comment": {
                "properties": {
                    "content": { "type": "string", "title": "Content" },
                    "commenter_id": { "type": "integer", "title": "Commenter Id" },
                    "post_id": { "type": "integer", "title": "Post Id" },
                    "id": { "type": "integer", "title": "Id" },
                    "posted_on": { "type": "string", "format": "date-time", "title": "Posted On" },
                    "edited": { "type": "boolean", "title": "Edited" },
                    "commenter": { "type": "string", "title": "Commenter" }
                },
                "type": "object",
                "required": [
                    "content",
                    "commenter_id",
                    "post_id",
                    "id",
                    "posted_on",
                    "edited",
                    "commenter"
                ],
                "title": "Comment"
            },
            "CommentCreate": {
                "properties": {
                    "content": { "type": "string", "title": "Content" },
                    "commenter_id": { "type": "integer", "title": "Commenter Id" },
                    "post_id": { "type": "integer", "title": "Post Id" }
                },
                "type": "object",
                "required": ["content", "commenter_id", "post_id"],
                "title": "CommentCreate"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": { "$ref": "#/components/schemas/ValidationError" },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Like": {
                "properties": {
                    "user_id": { "type": "integer", "title": "User Id" },
                    "post_id": { "type": "integer", "title": "Post Id" },
                    "id": { "type": "integer", "title": "Id" },
                    "liked_on": { "type": "string", "format": "date-time", "title": "Liked On" },
                    "user": { "type": "string", "title": "User" }
                },
                "type": "object",
                "required": ["user_id", "post_id", "id", "liked_on", "user"],
                "title": "Like"
            },
            "Post": {
                "properties": {
                    "content": { "type": "string", "title": "Content" },
                    "poster_id": { "type": "integer", "title": "Poster Id" },
                    "id": { "type": "integer", "title": "Id" },
                    "posted_on": { "type": "string", "format": "date-time", "title": "Posted On" },
                    "edited": { "type": "boolean", "title": "Edited" },
                    "comments": {
                        "anyOf": [
                            {
                                "items": { "$ref": "#/components/schemas/Comment" },
                                "type": "array"
                            },
                            { "type": "null" }
                        ],
                        "title": "Comments"
                    },
                    "likes": {
                        "anyOf": [
                            { "items": { "$ref": "#/components/schemas/Like" }, "type": "array" },
                            { "type": "null" }
                        ],
                        "title": "Likes"
                    }
                },
                "type": "object",
                "required": ["content", "poster_id", "id", "posted_on", "edited"],
                "title": "Post"
            },
            "PostCreate": {
                "properties": {
                    "content": { "type": "string", "title": "Content" },
                    "poster_id": { "type": "integer", "title": "Poster Id" }
                },
                "type": "object",
                "required": ["content", "poster_id"],
                "title": "PostCreate"
            },
            "User": {
                "properties": {
                    "username": { "type": "string", "title": "Username" },
                    "id": { "type": "integer", "title": "Id" },
                    "signed_up_on": {
                        "anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }],
                        "title": "Signed Up On"
                    },
                    "posts": {
                        "anyOf": [
                            { "items": { "$ref": "#/components/schemas/Post" }, "type": "array" },
                            { "type": "null" }
                        ],
                        "title": "Posts"
                    },
                    "comments": {
                        "anyOf": [
                            {
                                "items": { "$ref": "#/components/schemas/Comment" },
                                "type": "array"
                            },
                            { "type": "null" }
                        ],
                        "title": "Comments"
                    },
                    "likes": {
                        "anyOf": [
                            { "items": { "$ref": "#/components/schemas/Like" }, "type": "array" },
                            { "type": "null" }
                        ],
                        "title": "Likes"
                    }
                },
                "type": "object",
                "required": ["username", "id"],
                "title": "User"
            },
            "UserCreate": {
                "properties": {
                    "username": { "type": "string", "title": "Username" },
                    "password": { "type": "string", "title": "Password" }
                },
                "type": "object",
                "required": ["username", "password"],
                "title": "UserCreate"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": { "type": "string", "title": "Message" },
                    "type": { "type": "string", "title": "Error Type" }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": { "password": { "scopes": {}, "tokenUrl": "/auth/sign-in" } }
            }
        }
    }
}